scripts/requirements.txt
rdflib>=7.0.0
pyshacl>=0.25.0
graphviz>=0.20.1
PyYAML>=6.0.0
python-dotenv>=1.0.0